fix(grovedb,query): address review feedback on carrier ACOR
- Move ACOR construction/validation out of `grovedb-query/src/query.rs`
  (it had grown past 1500 lines) into a focused
  `grovedb-query/src/aggregate_count.rs` with its own `impl Query` block
  and tests. `query.rs` shrinks from ~1580 to ~910 lines.

- Per CodeRabbit major: route verifier validation through the
  PathQuery layer so `SizedQuery::limit`/`offset` are rejected up
  front for both leaf and carrier shapes. New tests pin the
  pagination rejection.

- V0 proof envelopes can't carry the carrier shape (they're produced
  only by pre-feature grove versions); the new `verify_v0_with_classification`
  rejects carrier classification with `InvalidProof("only supported on V1
  proof envelopes")` instead of attempting a never-emitted carrier walk.
  Removes the unused `verify_v0_carrier_layer` / `verify_v0_subquery_path`.

- Per CodeRabbit minor: ordering doc on
  `verify_aggregate_count_query_per_key` now says "query-direction
  order" (matches the merk walker) instead of asserting strict
  ascending lex, since the carrier honors `Query::left_to_right`.

- Per CodeRabbit nitpick: pin the malformed-leaf-inner-Key rejection
  message in the carrier path so future refactors can't silently
  re-route the error.

- New end-to-end tests: V0+carrier rejection, long subquery_path
  (2 intermediate single-key layers), corrupted outer-layer merk
  bytes, undecodable envelope, legacy `verify_aggregate_count_query`
  rejection of carrier shapes, carrier pagination rejection.
